tests: Fix and improve several aspects of existing tests
