+++ tests/Flac/FlacTaggingTest.php
@@ -28,22 +28,41 @@ class FlacTaggingTest extends TestCase
      */
     public function testFlacTaggerCanTagFlacFile()
     {
-        $this->assertIsArray(
-            $this->flacManipulator->tagger->writeTags(
-                $this->flacManipulator->getFile(),
-                [
-                    'songId' => 1,
-                    'name' => 'Test Song',
-                    'songOrder' => 1,
-                    'trackPreviewStart' => 0,
-                    'moniker' => 'Foobius Barius',
-                    'title' => 'Test Title',
-                    'year' => '1981',
-                    'genre' => 1,
-                    'albumId' => 1,
-                    'license' => 'All rights reserved.',
-                ]
-            )
+        $tagData = [
+            'songId' => 1,
+            'name' => 'Test Song',
+            'songOrder' => 1,
+            'trackPreviewStart' => 0,
+            'moniker' => 'Foobius Barius',
+            'title' => 'Test Title',
+            'year' => '1981',
+            'genre' => 1,
+            'albumId' => 1,
+            'license' => 'All rights reserved.',
+        ];
+
+        $this->flacManipulator->tagger->writeTags(
+            $this->flacManipulator->getFile(),
+            $tagData
+        );
+
+        $fileDetails = $this->flacManipulator
+            ->tagger
+            ->getid3
+            ->analyze($this->flacManipulator->getFile());
+
+        $this->assertEquals(
+            [
+                'title' => [$tagData['name']],
+                'artist' => [$tagData['moniker']],
+                'date' => [$tagData['year']],
+                'description' => [$tagData['license']],
+                'album' => [$tagData['title']],
+                'discnumber' => ['1/1'],
+                'tracknumber' => [$tagData['songOrder'] . '/1'],
+                'genre' => ['Rock'],
+            ],
+            $fileDetails['tags']['vorbiscomment']
         );
     }
 }
